Корректировка кода согласно замечаниям ревьювера
